xsdatetime fixes, broke datetime functions out
