+++ codalib/bagatom.py
@@ -3,6 +3,7 @@ from . import anvl, APP_AUTHOR
 import urllib
 from lxml import etree
 from datetime import datetime
+from codalib.xsdatetime import xsDateTime_format
 
 TIME_FORMAT_STRING = "%Y-%m-%dT%H:%M:%SZ"
 DATE_FORMAT_STRING = "%Y-%m-%d"
@@ -42,9 +43,9 @@ def wrapAtom(xml, id, title, author=None, updated=None, author_uri=None,
             type=alt_type)
 
     if updated is not None:
-        updatedTag.text = updated.strftime(TIME_FORMAT_STRING)
+        updatedTag.text = xsDateTime_format(updated)
     else:
-        updatedTag.text = datetime.utcnow().strftime(TIME_FORMAT_STRING)
+        updatedTag.text = xsDateTime_format(datetime.now())
     if author or author_uri:
         authorTag = etree.SubElement(entryTag, ATOM + "author")
         if author:
@@ -295,7 +296,7 @@ def makeObjectFeed(
         urlTag.text = author.get('uri', 'http://library.unt.edu/')
     # the updated tag is a
     updatedTag = etree.SubElement(feedTag, ATOM + "updated")
-    updatedTag.text = datetime.utcnow().strftime(TIME_FORMAT_STRING)
+    updatedTag.text = xsDateTime_format(datetime.now())
     # we will always show the link to the current 'self' page
     linkTag = etree.SubElement(feedTag, ATOM + "link")
     linkTag.set("rel", "self")
